Edit Scaffold
- updated the relationships in the model, serializer
- updated controller with all actions. in the create method bound the
profile role to a variable and then made a conditional to check
that the profile role was kitchen_owner before it creates something.
- updated data table and removed all the individual equipment column
and made it a description field that you can just type in.
- added seed data
- wrote create/index/update scripts for kitchens

At this point quite a few things are not working with the controller
and the scripts.
Next I will revert back to the profile and remove the role column